requirements.txt
openai
wikipedia-api
google-cloud-videointelligence
